+++ src/tao/tutorials/ex2.c
@@ -283,11 +283,11 @@ int main(int argc, char **argv)
     PetscCall(VecViewFromOptions(u, NULL, "-initial_vec_view"));
     PetscCall(DMComputeL2Diff(dm, 0.0, exactFuncs, NULL, u, &error));
     if (error < 1.0e-11) PetscCall(PetscPrintf(PETSC_COMM_WORLD, "Initial L_2 Error: < 1.0e-11\n"));
-    else                 PetscCall(PetscPrintf(PETSC_COMM_WORLD, "Initial L_2 Error: %g\n", error));
+    else                 PetscCall(PetscPrintf(PETSC_COMM_WORLD, "Initial L_2 Error: %g\n", (double)error));
     PetscCall(SNESSolve(snes, NULL, u));
     PetscCall(DMComputeL2Diff(dm, 0.0, exactFuncs, NULL, u, &error));
     if (error < 1.0e-11) PetscCall(PetscPrintf(PETSC_COMM_WORLD, "Final L_2 Error: < 1.0e-11\n"));
-    else                 PetscCall(PetscPrintf(PETSC_COMM_WORLD, "Final L_2 Error: %g\n", error));
+    else                 PetscCall(PetscPrintf(PETSC_COMM_WORLD, "Final L_2 Error: %g\n", (double)error));
   }
   PetscCall(VecViewFromOptions(u, NULL, "-sol_vec_view"));
